+++ api/fe.js
@@ -1,0 +1,3 @@
+const { generateContainerScript } = require('../forward_engineering/helpers/generateContainerScript');
+
+module.exports = { generateContainerScript };
